use a Rust nightly compiler
For building an OS we will need some experimental features that are
only available on the nightly channel, so we need to install a
nightly version of Rust.

We're using a Rust nightly compiler on this specific project only.
So, we need to add a file called `rust-toolchain` to the project’s root
directory.