handle VGA text buffer newlines
Right now, we just ignore newlines and characters that don’t fit into
the line anymore. Instead we want to move every character one line
up (the top line gets deleted) and start at the beginning of the last
line again. To do this, we add an implementation for the
`new_line` method of `Writer`.