spinlocks in Rust with #[no_std]
To get synchronized interior mutability, users of the standard
library can use Mutex. It provides mutual exclusion by blocking
threads when the resource is already locked. But our basic kernel
does not have any blocking support or even a concept of threads, so
we can’t use it either. However there is a really basic kind of mutex
in computer science that requires no operating system features: the
spinlock. Instead of blocking, the threads simply try to lock it
again and again in a tight loop and thus burn CPU time until the
mutex is free again.

To use a spinning mutex, we add the `spin` crate as a dependency.