make cargo test to exit after running all tests
Right now we have an endless loop at the end of our `_start` function
and need to close QEMU manually on each execution of cargo test.

The clean solution to this would be to implement a proper way to
shutdown our OS. Unfortunately this is relatively complex.

Luckily, there is an escape hatch: enable QEMU special
`isa-debug-exit` device, which provides an easy way to exit QEMU from
the guest system. We do so by adding a configuration key in
our `Cargo.toml`.