fix: success exit code
The problem is that `cargo test` interprets the test as failed even
though we passed our `Success` exit code.

The problem is that `cargo test` considers all error codes other than
`0` as failure.

To work around this, `bootimage` provides a `test-success-exit-code`
config key that maps a specified exit code to the exit code `0`.
With this config, `cargo test` correctly recognizes the success case
and does not count the test as failed.

Our test runner now automatically closes QEMU and correctly reports
the test results out.