print to the console, serial port module
We still see the QEMU window open for a very short time, but it does
not suffice to read the results. It would be nice if we could print
the test results to the console instead, so that we can still see
them after QEMU exited.

To see the test output on the console, we need to send the data from
our kernel to the host system somehow.

A simple way to send the data is to use the serial port.
QEMU can redirect the bytes sent over serial to the host’s standard
output or a file.

Create a new serial module and use the `uart_16550` crate to
initialize the UART and send data over the serial port.

To see the serial output from QEMU, we need use the `-serial`
argument to redirect the output to stdout.