switching stacks & avoid stack overflow problem
Now the CPU should switch to the double fault stack whenever a double
fault occurs. Thus, we are able to catch all double faults,
including kernel stack overflows.

Before we implement the final steps:
we still see the boot loop on stack overflow.

The problem is that the GDT segments are not yet active because the
segment and TSS registers still contain the values from the old GDT.
We also need to modify the double fault IDT entry so that it uses
the new stack.

We did the following:
* Reload code segment register.
* Load the TSS.
* Update the IDT entry.