end of interrupt
In our timer interrupt handler, we print a dot to the screen. As the
timer interrupt happens periodically, we would expect to see a dot
appearing on each timer tick. However, when we run it we see that
only a single dot is printed.

The reason is that the PIC expects an explicit
"end of interrupt" (EOI) signal from our interrupt handler.
This change send the EOI.

When we now execute `cargo run` we see dots periodically appearing
on the screen.