provoke a deadlock
When we run it in QEMU, we see that only a limited number of hyphens
is printed, until the first timer interrupt occurs. Then the system
hangs because the timer interrupt handler deadlocks when it tries to
print a dot. This is the reason that we see no dots in the output.