fix a race condition
If you run `cargo test` you might see the `test_println_output` test
failing. The reason is a race condition between the test and our
timer handler.

The race condition occurs because the timer interrupt handler might
run between the `println` and the reading of the screen characters.

To fix this, we keep the `WRITER` locked for the complete duration of
the test, so that the timer handler can't write a `.` to the screen
in between.