efficient endless loop using hlt instruction
Until now we used a simple empty loop statement at the end of our
`_start` and `panic` functions. This causes the CPU to spin
endlessly and thus works as expected. But it is also very
inefficient, because the CPU continues to run at full speed even
though there’s no work to do.

What we really want to do is to halt the CPU until the next
interrupt arrives. This allows the CPU to enter a sleep state in
which it consumes much less energy. The `hlt` instruction does
exactly that. We use this instruction to create an energy efficient
endless loop.

When we run our kernel now in QEMU, we see a much lower CPU usage.