add support for keyboard input
Keyboard input will allow us to interact with our kernel for the
first time.

We add a handler function for the keyboard interrupt. It's quite
similar to how we defined the handler for the timer interrupt, it
just uses a different interrupt number.

We now see that a `k` appears on the screen when we press a key.