translate the other keys
Translating the other keys works in the same way. Fortunately there
is a crate named `pc-keyboard` for translating scancodes of scancode
sets 1 and 2, so we don't have to implement this ourselves.

We use this crate to rewrite our `keyboard_interrupt_handler`.

With this modified interrupt handler we can now write text.