memory management: page faults
Try to cause a page fault by accessing some memory outside of our
kernel. First, we create a page fault handler and register it in our
IDT, so that we see a page fault exception instead of a generic
double fault.

Then we can try to access some memory outside our kernel.

When we run it, we see that our page fault handler is called.