enable bootloader support for accessing page tables
To make the page table frames accessible to kernel,
we choose the approach that maps the complete physical memory.

To implement the approach, we will need support from the bootloader,
so we configure it first.

To enable the required bootloader support, we add the
`map_physical_memory` feature to our `bootloader` dependency.

With this feature enabled, the bootloader maps the complete physical
memory to some unused virtual address range. To communicate the
virtual address range to our kernel, the bootloader passes a boot
information structure.

The `bootloader` crate defines a `BootInfo` struct that contains all
the information it passes to our kernel.
With the `map_physical_memory` feature enabled, it currently has the
two fields `memory_map` and `physical_memory_offset`.