accessing the page tables
Now that we have access to physical memory, we can finally start to
implement our page table code.

Before we begin, we create a new `memory` module for our code.

We create an `active_level_4_table` function that returns a reference
to the active level 4 page table. Then, we use this function to print
the entries of the level 4 table.

We traverse the page tables further and take a look at a
level 3 table.