translating addresses
We create a translation function that returns the physical address
that a given virtual address is mapped to.

For translating a virtual to a physical address, we have to traverse
the four-level page table until we reach the mapped frame. We create
a function that performs this translation.

We test our translation function by translating some addresses
in `kernel_main()`.