using OffsetPageTable
Translating virtual to physical addresses is a common task in an OS
kernel, therefore the `x86_64` crate provides an abstraction for it.

The `OffsetPageTable` type assumes that the complete physical memory
is mapped to the virtual address space at some offset.

In our case, the bootloader maps the complete physical memory at a
virtual address specified by the `physical_memory_offset` variable,
so we can use the `OffsetPageTable` type. To initialize it, we create
a new `init` function in our `memory` module.

We now can use the `Translate::translate_addr` method instead of our
own `memory::translate_addr` function. We only need to change a few
lines in our `kernel_main`.

When we run it now, we see the same translation results as before,
with the difference that the huge page translation now also works.