create a new mapping in the page tables
The first step of our implementation is to create a new
`create_example_mapping` function that maps a given virtual page
to `0xb8000`, the physical frame of the VGA text buffer.

To be able to call `create_example_mapping` we need to create a type
that implements the `FrameAllocator` trait first.

We modify our `kernel_main` function to map the page at virtual
address `0`. Since we map the page to the frame of the VGA text
buffer, we are able to write to the screen through it.