allocating frames
In order to create new page tables, we need to create a proper frame
allocator. For that we use the `memory_map` that is passed by the
bootloader as part of the `BootInfo` struct.

Before we implement the `FrameAllocator` trait, we add an auxiliary
`usable_frames` method  that converts the memory map into an
iterator of usable frames.

We use the `BootInfoFrameAllocator` in `main.rs`. We modify
our `kernel_main` function to pass a `BootInfoFrameAllocator`
instance instead of an `EmptyFrameAllocator`.