add dependency on the built-in alloc crate
We will create a heap memory region for our kernel, which will allow
us to allocate memory and use various collection types.

The first step in implementing a heap allocator is to add a
dependency on the built-in `alloc` crate.

Since we are compiling for a custom target, we can’t use the
precompiled version of `alloc` that is shipped with the Rust
installation. Instead, we have to tell cargo to recompile the crate
from source.