create a kernel heap
Before we can create a proper allocator, we first need to create a
heap memory region from which the allocator can allocate memory.
To do this, we need to define a virtual memory range for the heap
region and then map this region to physical frames.
We create the `init_heap` function for this.

We now have a mapped heap memory region that is ready to be used.