use an allocator crate
Since implementing an allocator is somewhat complex, we start by
using an external allocator crate. We will implement our own
allocator later.

To use the `alloc` crate, we add a dependency on it in Cargo.

Then, we replace our dummy allocator with the allocator provided by
the crate.

Setting the `LockedHeap` as global allocator is not enough.
We need to initialize the allocator after creating the heap.

After initializing the heap, we can now use all allocation and
collection types of the built-in `alloc` crate.