use the linked list allocator
We update the `ALLOCATOR` static in the `allocator` module to use our
new `LinkedListAllocator`.

When we run our `heap_allocation` tests again, we see that all tests
pass now, including the `many_boxes_long_lived` test that failed with
the bump allocator.

This shows that our linked list allocator is able to reuse freed
memory for subsequent allocations.