fix: race condition in sleep_if_idle
There is still a subtle race condition in this implementation. Since
interrupts are asynchronous and can happen at any time, it is
possible that an interrupt happens right between the
`task_queue.is_empty` check and the call to HLT instruction.

In case this interrupt pushes to the `task_queue`, we put the CPU to
sleep even though there is now a ready task.

To prevent it, we disable interrupts on the CPU before the check and
atomically enable them again together with the HLT instruction.
This way, all interrupts that happen in between are delayed after the
HLT instruction so that no wake-ups are missed.

To fix this, we update the implementation of `sleep_if_idle` function.

Now our executor properly puts the CPU to sleep when there is
nothing to do. We can see that the QEMU process has a much lower CPU
utilization when we run our kernel again.